Update profile entity
